To avoid complexity, I made JMP & CALL operand size static to 32bits

+++ refs.c
@@ -37,17 +37,17 @@ void resolve_refs() {
     // }
     bool found;
     uint64_t found_addr;
-    uint8_t relative_addr;
+    uint32_t relative_addr;
     for (uint64_t i = 0; i < unresolved_refrences_size; i++) {
         found_addr = search_refs(unresolved_refrences[i].name, unresolved_refrences[i].type, &found);
         if (found) {
             if (unresolved_refrences[i].type == JMP_REF) {
-                relative_addr = (uint8_t) refrences[found_addr].addr - (unresolved_refrences[i].addr-1);
-                memcpy(&_functions[unresolved_refrences[i].func_idx].code[unresolved_refrences[i].addr], &relative_addr, sizeof(int8_t));
+                relative_addr = (uint32_t) refrences[found_addr].addr - (unresolved_refrences[i].addr-1);
+                memcpy(&_functions[unresolved_refrences[i].func_idx].code[unresolved_refrences[i].addr], &relative_addr, sizeof(uint32_t));
             }
             else if (unresolved_refrences[i].type == FUNC_REF) {
-                relative_addr = (int8_t) refrences[found_addr].func_idx;
-                memcpy(&_functions[unresolved_refrences[i].func_idx].code[unresolved_refrences[i].addr], &relative_addr, sizeof(int8_t));
+                relative_addr = (uint32_t) refrences[found_addr].func_idx;
+                memcpy(&_functions[unresolved_refrences[i].func_idx].code[unresolved_refrences[i].addr], &relative_addr, sizeof(uint32_t));
             }
             else {
                 printf("unknown ref typen\n"); exit(1);
